requirements.txt
# Core dependencies
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI and NLP
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
chromadb==0.4.22
tiktoken==0.5.2

# Document processing
python-docx==1.1.0
pandas==2.2.0
numpy>=1.26.0
openpyxl==3.1.2

# PDF generation
reportlab==4.0.8
weasyprint==60.2
jinja2==3.1.3

# Database and storage
sqlalchemy==2.0.25
redis==5.0.1

# HTTP and API
httpx~=0.25.2  # Compatible with python-telegram-bot 20.7
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
pendulum==3.0.0

# Logging and monitoring
structlog==24.1.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
ruff==0.1.11
mypy==1.8.0
pre-commit==3.6.0

# Type stubs
types-requests==2.31.0.20240106
types-aiofiles==23.2.0.20240106